+++ packages/commands/changelog.md
@@ -1,9 +1,27 @@
-#  (2025-04-17)
+#  (2025-04-23)
 
 
 ### Bug Fixes
 
+* bindings implementation 8362190
+* ensure InitAkala warms up correctly when handling the init command c7dd7a2
 * extract params considers when array is provided c5eb98b
+* force run init-akala only once (no matter the trigger) 1acbe99
+* improve resolvable wrintings 6be2301
+* metadata generation when provided invalid names b9b828e
+* metatadata to return original commands 27a9049
+* unit tests ddef0ab
+* update after core breaking change 2eae6b4
+* update error status code for unauthorized actions in AuthHandler 663f8b9
+
+
+### Features
+
+* add auth to all configurations f48da73
+* add bindings (as in azure function bindings) as configuration 69067b9
+* add bindings (as in azure function bindings) as configuration 12850c1
+* add http auth in schema and auth in general 4cf5e55
+* add InitAkala command processor 652270a