+++ packages/config/changelog.md
@@ -1,4 +1,9 @@
-#  (2025-04-17)
+#  (2025-04-23)
+
+
+### Bug Fixes
+
+* broken `has` behavior de026d0